+++ unity/Assets/Scripts/Interfaces/ITileType.cs
@@ -1,6 +1,6 @@
 ﻿using Hexxle.TileSystem;
 
-namespace Assets.Scripts.Interfaces
+namespace Hexxle.Interfaces
 {
     public interface ITileType
     {
