+++ unity/Assets/Scripts/TileSystem/ENature.cs
@@ -1,4 +1,4 @@
-﻿namespace Assets.Scripts.TileSystem
+﻿namespace Hexxle.TileSystem
 {
     public enum ENature
     {
