+++ unity/Assets/Scripts/TileSystem/Tile.cs
@@ -1,7 +1,7 @@
-using Assets.Scripts.Interfaces;
-using Assets.Scripts.TileSystem;
-using Assets.Scripts.TileSystem.Nature;
-using Assets.Scripts.TileSystem.Type;
+using Hexxle.Interfaces;
+using Hexxle.TileSystem.Nature;
+using Hexxle.TileSystem.Behaviour;
+using Hexxle.TileSystem.Type;
 using Hexxle.CoordinateSystem;
 
 namespace Hexxle.TileSystem
