+++ unity/Assets/Scripts/Unity/UnityUI.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+using UnityEngine.SceneManagement;
+
+namespace Hexxle.Unity
+{
+    public class UnityUI : MonoBehaviour
+    {
+        public void exitRound()
+        {
+            SceneManager.LoadScene("titlescreen");
+        }
+    }
+}
