--- unity/Assets/Scripts/UnityMonoBehaviour/UnityUI.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class UnityUI : MonoBehaviour
-{
-    public void exitRound()
-    {
-        SceneManager.LoadScene("titlescreen");
-    }
-}
